+++ gateway/go.mod
@@ -1,0 +1,19 @@
+module github.com/cahllagerfeld/go-service-v2/gateway
+
+go 1.17
+
+require (
+	github.com/cahllagerfeld/go-service-v2/album-service v0.0.0-20220314074005-e95bfdaeaf7f
+	github.com/cahllagerfeld/go-service-v2/number-service v0.0.0-20220314074005-e95bfdaeaf7f
+	github.com/gorilla/mux v1.8.0
+	google.golang.org/grpc v1.45.0
+)
+
+require (
+	github.com/golang/protobuf v1.5.2 // indirect
+	golang.org/x/net v0.0.0-20220225172249-27dd8689420f // indirect
+	golang.org/x/sys v0.0.0-20220310020820-b874c991c1a5 // indirect
+	golang.org/x/text v0.3.7 // indirect
+	google.golang.org/genproto v0.0.0-20220314164441-57ef72a4c106 // indirect
+	google.golang.org/protobuf v1.27.1 // indirect
+)
